twitter_authentication_and_handshake. R
The twitter_authentication_and_handshake. R is used to establish connection with twitter. Initially an application was created using the dev.twitter.com for performing the sentiment analysis and the credentials to access the application are extracted. Thus the extracted keys and token are used in the handshake and authentication process.